move readdirRecursive to helpers.js
